dosen n mahasiswa -> user
gabung dosen mahasiswa mc to user mc
