--- ta-app/app/Http/Controllers/MahasiswaController.php
@@ -1,68 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\Mahasiswa;
-use Illuminate\Http\Request;
-
-class MahasiswaController extends Controller
-{
-    public function index()
-    {
-        $mahasiswas = Mahasiswa::all();
-        return view('mahasiswa.index', compact('mahasiswas'));
-    }
-
-    public function create()
-    {
-        return view('mahasiswa.create');
-    }
-
-    public function store(Request $request)
-    {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|email|unique:users',
-            'roles' => 'required|string|max:255',
-            'jurusan' => 'required|string|max:255', // Validasi untuk jurusan
-            'password' => 'required|string|min:8',
-        ]);
-
-        $data = $request->all();
-        $data['password'] = bcrypt($request->password);
-        Mahasiswa::create($data);
-
-        return redirect()->route('mahasiswa.index')->with('success', 'Mahasiswa berhasil ditambahkan!');
-    }
-
-    public function show(Mahasiswa $mahasiswa)
-    {
-        return view('mahasiswa.show', compact('mahasiswa'));
-    }
-
-    public function edit(Mahasiswa $mahasiswa)
-    {
-        return view('mahasiswa.edit', compact('mahasiswa'));
-    }
-
-    public function update(Request $request, Mahasiswa $mahasiswa)
-    {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|email|unique:users,email,' . $mahasiswa->id,
-            'roles' => 'required|string|max:255',
-            'jurusan' => 'required|string|max:255', // Validasi untuk jurusan
-        ]);
-
-        $mahasiswa->update($request->all());
-
-        return redirect()->route('mahasiswa.index')->with('success', 'Data mahasiswa berhasil diperbarui!');
-    }
-
-    public function destroy(Mahasiswa $mahasiswa)
-    {
-        $mahasiswa->delete();
-
-        return redirect()->route('mahasiswa.index')->with('success', 'Mahasiswa berhasil dihapus!');
-    }
-}
