--- ta-app/app/Models/Dosen.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Dosen extends Model
-{
-    //
-}
