+++ ta-app/resources/views/main/index.blade.php
@@ -1,43 +1,5 @@
-@extends('layouts.app')
-
-@section('content')
-<div class="container">
-    <h1>Daftar Mahasiswa</h1>
-    <a href="{{ route('mahasiswa.create') }}" class="btn btn-primary">Tambah Mahasiswa</a>
-
-    @if (session('success'))
-        <div class="alert alert-success">{{ session('success') }}</div>
-    @endif
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>ID</th>
-                <th>Nama</th>
-                <th>Email</th>
-                <th>Roles</th>
-                <th>Aksi</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach ($mahasiswas as $mahasiswa)
-                <tr>
-                    <td>{{ $mahasiswa->id }}</td>
-                    <td>{{ $mahasiswa->name }}</td>
-                    <td>{{ $mahasiswa->email }}</td>
-                    <td>{{ $mahasiswa->roles }}</td>
-                    <td>
-                        <a href="{{ route('mahasiswa.show', $mahasiswa->id) }}" class="btn btn-info">Detail</a>
-                        <a href="{{ route('mahasiswa.edit', $mahasiswa->id) }}" class="btn btn-warning">Edit</a>
-                        <form action="{{ route('mahasiswa.destroy', $mahasiswa->id) }}" method="POST" style="display:inline-block;">
-                            @csrf
-                            @method('DELETE')
-                            <button type="submit" class="btn btn-danger">Hapus</button>
-                        </form>
-                    </td>
-                </tr>
-            @endforeach
-        </tbody>
-    </table>
-</div>
-@endsection
+<html>
+    <body>
+        <h1>Hello World</h1>
+    </body>
+</html>
